+++ example/services/auth.dart
@@ -3,17 +3,17 @@ import 'dart:io';
 
 import 'package:lite_server/lite_server.dart';
 
-class AuthService extends HttpService {
+class AuthService extends HttpController {
   @override
-  FutureOr<HttpServiceBehavior> onRequest(HttpRequest request) {
-    return HttpServiceBehavior.next(extra: {'auth': 1});
+  FutureOr<HttpControllerBehavior> onRequest(HttpRequest request) {
+    return HttpControllerBehavior.next(extra: {'auth': 1});
   }
 }
 
-class IpCheckService extends HttpService {
+class IpCheckService extends HttpController {
   @override
-  FutureOr<HttpServiceBehavior> onRequest(HttpRequest request) {
+  FutureOr<HttpControllerBehavior> onRequest(HttpRequest request) {
     request.response.unauthorized();
-    return HttpServiceBehavior.revoke();
+    return HttpControllerBehavior.revoke();
   }
 }